Initial (Work-in-Progress) Tweed RunC support (#2)
This commit introduces a work-in-progress port of Tweed stencils from
assets-on-disk (executing in the same namespaces as the broker process
itself) to OCI images that use an exec-style protocol for activating life
cycle hooks.

Images are loaded into a local registry at boot time, from a static
configuration (pulled from default registries like other container-based
tooling).  This is done asynchronously and in parallel to decrease startup
times.