Expect error before inifnite loop
Previously, if the `registry` binary is not installed (i.e. you try to
run the unit tests outside of the testing container), the tests would
hang until they timed out, because the error assertion about executing
the binary wouldn't get checked until _after_ the sleep loop waiting for
the registry to boot.

This has been fixed, and now we validate that the registry command
executed in the background _before_ we sleep-loop trying to connect to
it.