[osb] Only Inflate Instance Params When Needed
This fixes an issue where instance parameters are not sent, and rather
than getting an empty map (`{}`), we get nil, which fails to JSON
unmarshal, which casuses the broker to choke, which makes the CF app
user person unhappy.  We don't want our CF app user people to be
unhappy, now do we?

Fixes #19